Renamed IterationEvent EVENT_ITERATE* constants to EVENT_ITERATION*.
In the future it might be nice to have events before and after the
foreach that iterates through the instances. These events likely would
be intuitively called pre-iterate and post-iterate events, which would
conflict with the events within the foreach. The rename of the events
within the foreach from iterate to iteration makes the event names more
intuitive and opens the door to the aforementioned new events in the
future.